+++ server/src/server.ts
@@ -172,7 +172,7 @@ class CSSNavigationServer {
 			}
 		})
 
-		Logger.log(`CSS Navigation Service for workspace "${path.basename(this.options.workspaceFolderPath)}" started.`)
+		Logger.log(`💼 CSS Navigation Service for workspace "${path.basename(this.options.workspaceFolderPath)}" started.`)
 	}
 
 	private updateTimestamp(time: number) {