+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 - Removes `ignoreSameNameCSSFile` option, if you are using it, use `excludeGlobPatterns` or `ignoreFilesBy` instead.
 - Fix the issue that can't find references from not opened js or ts files.
-- Can rightly release those not been included in workspace sources after they get closed.
+- Can rightly release those not been included in workspace resources after they get closed.
 - Will release long-unused resources, and reload them again if required.
 - Can follow import chain when searching definitions.